+++ src/modules/newsletter-form/styles.module.css
@@ -1,0 +1,75 @@
+.subscribeForm {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+.newsletterInput {
+  color: var(--ifm-color-black);
+  font-family: "Roboto", sans-serif;
+  font-size: 1rem;
+  width: 100%;
+  max-width: 300px;
+  min-width: 100px;
+  background: var(--ifm-color-white);
+  border: 1px solid var(--ifm-color-white-darkest);
+  box-sizing: border-box;
+  box-shadow: rgba(0, 0, 0, 0.05) 0px 1px 2px;
+  border-radius: 6px;
+  padding: .4rem 0.75rem;
+  margin: 0 .625rem 0 0 !important;
+}
+.successMessage {
+  font-family: "Inter", sans-serif;
+  color: var(--ifm-color-white-dark);
+  font-size: 1rem;
+}
+.submitBtn {
+  background: #b65037;
+  font-size: 1rem;
+  color: var(--ifm-color-white);
+  font-family: "Inter", sans-serif;
+  max-width: 300px;
+  width: 100%;
+  white-space: normal;
+  height: 38px;
+  align-items: center;
+  justify-content: center;
+  flex-direction: row;
+  padding: 9px 17px;
+  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.05);
+  border-radius: 6px;
+  text-align: center;
+  font-style: normal;
+  font-weight: 500;
+  line-height: var(--ifm-font-line-height);
+  border: none;
+  cursor: pointer;
+  white-space: nowrap !important;
+}
+.backBtn {
+  color: var(--ifm-color-gray-light);
+  font-size: 1rem;
+  margin: .625rem auto;
+  text-align: center;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  text-decoration: none;
+}
+.backBtn:hover {
+  text-decoration: underline;
+}
+.errorMessage {
+  color: rgb(185, 28, 28);
+  font-size: 1rem;
+}
+@media (max-width: 768px) {
+  .subscribeForm {
+    flex-direction: column;
+  }
+  .newsletterInput {
+    margin: .625rem 0 !important;
+  }
+}
