+++ src/modules/signup-form/styles.module.css
@@ -1,75 +1,68 @@
-.subscribeForm {
+.cta {
   display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
+  /* align-items: center; */
   width: 100%;
-}
-.newsletterInput {
   color: var(--ifm-color-black);
-  font-family: "Roboto", sans-serif;
-  font-size: 1rem;
-  width: 100%;
-  max-width: 300px;
-  min-width: 100px;
-  background: var(--ifm-color-white);
-  border: 1px solid var(--ifm-color-white-darkest);
-  box-sizing: border-box;
-  box-shadow: rgba(0, 0, 0, 0.05) 0px 1px 2px;
-  border-radius: 6px;
-  padding: .4rem 0.75rem;
-  margin: 0 .625rem 0 0 !important;
+  margin: 1rem 0 4.5rem;
 }
-.successMessage {
-  font-family: "Inter", sans-serif;
-  color: var(--ifm-color-white-dark);
-  font-size: 1rem;
-}
-.submitBtn {
-  background: #b65037;
-  font-size: 1rem;
-  color: var(--ifm-color-white);
-  font-family: "Inter", sans-serif;
-  max-width: 300px;
-  width: 100%;
-  white-space: normal;
-  height: 38px;
-  align-items: center;
-  justify-content: center;
-  flex-direction: row;
-  padding: 9px 17px;
-  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.05);
+.cta .contentWrapper {
+  background: var(--ifm-color-salmon-lightest);
   border-radius: 6px;
-  text-align: center;
+  padding: 3rem 3.3rem;
+}
+.cta .tagline {
+  font-family: var(--ifm-font-family-work-sans);
   font-style: normal;
-  font-weight: 500;
+  font-weight: 700;
+  font-size: 2rem;
+  width: 75%;
   line-height: var(--ifm-font-line-height);
-  border: none;
-  cursor: pointer;
-  white-space: nowrap !important;
+  letter-spacing: var(--ifm-font-letter-spacing);
+  margin-bottom: 1.5rem;
 }
-.backBtn {
-  color: var(--ifm-color-gray-light);
-  font-size: 1rem;
-  margin: .625rem auto;
-  text-align: center;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  text-decoration: none;
+.cta .desc {
+  font-family: var(--ifm-font-family-work-sans);
+  font-weight: 600;
+  font-size: 1.2rem;
+  line-height: var(--ifm-font-line-height);
+  letter-spacing: var(--ifm-font-letter-spacing);
 }
-.backBtn:hover {
-  text-decoration: underline;
+.cta .ctaBtn {
+  font-size: 1.1rem;
+  font-family: var(--ifm-font-family-work-sans);
+  padding: .2rem 1rem;
 }
-.errorMessage {
-  color: rgb(185, 28, 28);
-  font-size: 1rem;
-}
+
 @media (max-width: 768px) {
-  .subscribeForm {
-    flex-direction: column;
+  .cta {
+    margin: 2.5rem 0;
   }
-  .newsletterInput {
-    margin: .625rem 0 !important;
+  .cta .tagline {
+    font-size: 1.5rem;
+  }
+  .cta .contentWrapper {
+    padding: 1.8rem 2rem;
   }
 }
+@media (max-width: 1024px) {
+  .cta .tagline {
+    width: 100%;
+  }
+}
+
+    @media (max-width: 426px) {
+      .cta .tagline {
+        font-size: 1.05rem;
+        margin-bottom: 1rem;
+      }
+      .cta .desc {
+        font-size: 0.8rem;
+        margin-bottom: 1rem;
+      }
+      .cta .contentWrapper {
+        padding: 1.5rem 1rem;
+      }
+      .cta .ctaBtn {
+        font-size: 0.8rem;
+      }
+    }
