+++ src/modules/try-signoz-cta/styles.module.css
@@ -20,6 +20,7 @@
   text-align: center;
 }
 .cta .ctaBtn {
+  font-family: var(--ifm-font-family-work-sans);
   color: var(--ifm-color-primary) !important;
 }
 
@@ -30,8 +31,4 @@
   .cta .tagline {
     font-size: 1.5rem;
   }
-}
-
-html[data-theme="light"] .cta .tagline {
-  color: var(--ifm-color-white);
 }
